Remove the use of internal
