+++ error_handling/main.go
@@ -1,10 +1,48 @@
 package main
 
 import (
+	"encoding/json"
 	"errors"
+	"fmt"
 
-	"github.com/blocky/as-demo/internal"
+	"github.com/blocky/as-demo/as"
 )
+
+type Result struct {
+	Success bool
+	Error   string
+	Value   any
+}
+
+func (r Result) jsonMarshalWithError(err error) []byte {
+	if err == nil {
+		err = errors.New("jsonMarshalWithError invoked with nil error")
+	}
+	resultStr := fmt.Sprintf(
+		`{ "Success": false, "Error": "%v" , "Value": null }`,
+		err,
+	)
+	data := []byte(resultStr)
+	return data
+}
+
+func writeOutput(output any) uint64 {
+	result := Result{
+		Success: true,
+		Value:   output,
+	}
+	data, err := json.Marshal(result)
+	if err != nil {
+		as.Log(fmt.Sprintf("Error marshalling result: %v", err))
+		return writeError(err)
+	}
+	return as.WriteToHost(data)
+}
+
+func writeError(err error) uint64 {
+	data := Result{}.jsonMarshalWithError(err)
+	return as.WriteToHost(data)
+}
 
 //export successFunc
 func successFunc(inputPtr, secretPtr uint64) uint64 {
@@ -12,13 +50,13 @@ func successFunc(inputPtr, secretPtr uint64) uint64 {
 		Number int `json:"number"`
 	}
 	output := Output{42}
-	return internal.WriteOutput(output)
+	return writeOutput(output)
 }
 
 //export errorFunc
 func errorFunc(inputPtr, secretPtr uint64) uint64 {
 	err := errors.New("expected error")
-	return internal.WriteError(err)
+	return writeError(err)
 }
 
 func main() {}
